pyproject.toml
[tool.poetry]
name = "dragon-gpt"
version = "0.1.0"
description = ""
authors = ["Raja Giddi <97639175+RajaGiddi@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.2.8"
langchain-community = "^0.2.7"
python-dotenv = "^1.0.1"
langchain-core = "^0.2.20"
langchain-openai = "^0.1.16"
jq = "^1.7.0"
chromadb = "^0.5.4"
langchainhub = "^0.1.20"
langchain-chroma = "^0.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
